chore: Allow to build launchpad package from existing source pkg
This change allows to later add new dists for already uploaded
versions. This is necessary because the source package has to be
identical otherwise it'll be refused by the ppa.

It's necessary to download the source package (through
`apt source onboard-keyman` after enabling the `deb-src` keyman ppa)
to the parent directory of onboard-keyman. Afterwards the
`launchpad.sh` script can be run for the new dists.